Refactor spectroscopy classes to allow for variants
